refactor: tipos compartilhados, apiFetch com refresh, logout revoga sessão, Suspense no login
- utils/types.ts (NOVO): tipos de domínio centralizados (Status, Prioridade, Chamado, etc.)
- utils/api.ts: tenta refresh automático no 401 antes de deslogar; Content-Type
  não é sobrescrito para FormData (fix de upload)
- utils/auth.ts: logout chama POST /auth/logout para revogar sessão no banco
- login/LoginContent.tsx (NOVO): componente client extraído
- login/page.tsx: server component com Suspense — necessário para useSearchParams no Next.js 15"
